# backend/routers/auth.py
# generated by fastapi-codegen:
#   filename:  api.swagger.yaml
#   timestamp: 2024-04-18T22:51:52+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['auth'])


@router.get(
    '/auth/login',
    response_model=UserCredentials,
    responses={
        '401': {'model': Error},
        '403': {'model': Error},
        '500': {'model': Error},
    },
    tags=['auth'],
)
def get_auth_login() -> Union[UserCredentials, Error]:
    """
    Login user
    """
    return UserCredentials(access_token='access', refresh_token='<refresh>')


@router.post(
    '/auth/register',
    response_model=UserCredentials,
    responses={'400': {'model': Error}, '500': {'model': Error}},
    tags=['auth'],
)
def post_auth_register(body: User) -> Union[UserCredentials, Error]:
    """
    Register user
    """
    return Error(message='User is already registered')


@router.get(
    '/auth/token',
    response_model=UserCredentials,
    responses={
        '401': {'model': Error},
        '403': {'model': Error},
        '500': {'model': Error},
    },
    tags=['auth'],
)
def get_auth_token() -> Union[UserCredentials, Error]:
    """
    Refresh token
    """
    return UserCredentials(access_token='token', refresh_token='refresh')